+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 * A modified plugin is automatically saved before it is sent to a remote
   script.
 * Pyhton files can be dragged onto the editor to open them.
+* Annotation editor accessible through API.
 
 Version 0.3.0
 -------------